backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
numpy>=1.26.0
opencv-python-headless>=4.9.0
pillow>=10.3.0
# Optional: uncomment when adding real GeoTIFF support
# rasterio>=1.3.10
python-multipart
